+++ test/case/webpack-js-url/yyl.config.js
@@ -31,13 +31,13 @@ const setting = {
   proxy: {
     port: 8887,
     localRemote: {
-      'http://web.yy.com/': 'http://127.0.0.1:5000/'
+      'http://web.testhost.com/': 'http://127.0.0.1:5000/'
     },
-    homePage: `http://www.yy.com/web/${PROJECT_NAME}/`
+    homePage: `http://www.testhost.com/web/${PROJECT_NAME}/`
   }
 };
-setting.proxy.localRemote[`http://www.yy.com/web/${PROJECT_NAME}`] = `http://127.0.0.1:5000/project/${PROJECT_NAME}/${PLATFORM}/html`;
-setting.proxy.localRemote['http://www.yy.com/api/mock'] = 'http://127.0.0.1:5000/api/mock';
+setting.proxy.localRemote[`http://www.testhost.com/web/${PROJECT_NAME}`] = `http://127.0.0.1:5000/project/${PROJECT_NAME}/${PLATFORM}/html`;
+setting.proxy.localRemote['http://www.testhost.com/api/mock'] = 'http://127.0.0.1:5000/api/mock';
 // - setting
 
 const DEST_BASE_PATH = path.join(setting.localserver.root, setting.dest.basePath)
@@ -94,7 +94,7 @@ const config = {
     revAddr: `http://web.yystatic.com${setting.dest.basePath}/${setting.dest.revPath}/rev-manifest.json`,
     hostname: '//web.yystatic.com',
     staticHost: '//web.yystatic.com',
-    mainHost: '//www.yy.com/web'
+    mainHost: '//www.testhost.com/web'
   }
   // - configCommit
 }