+++ test/mocha/all.git.usercenter.test.js
@@ -1,3 +1,5 @@
 const { handleAllGit } = require('../fn/handle.all.git')
 
-handleAllGit('https://git.yy.com/webs/web_static/usercenter.git')
+handleAllGit(
+  'https://git*yy*com/webs/web_static/usercenter.git'.split('*').join('.')
+)