+++ test/mocha/all.git.yycom.test.js
@@ -1,3 +1,3 @@
 const { handleAllGit } = require('../fn/handle.all.git')
 
-handleAllGit('https://git.yy.com/webs/web_static/yycom.git')
+handleAllGit('https://git*yy*com/webs/web_static/yycom.git'.split('*').join('.'))